decomp: cleanup permuter output, remove debug logging
Source cleanup:
- Rename permuter-generated _tmp/_cond/_ref variables to readable names
  across 65 files (bgColor, teleported, nextShot, receiveEnd, etc.)
- Remove HX_NATIVE debug fprintf logging from CharClip, FileMerger,
  Flow, FlowSetProperty, DirLoader, Object, UIPanel, UIScreen, Dir.h
- Clean up ObjDirItr debug scaffolding in Dir.h (keep safety guards)
- Fix MoggClip::Play shadowing bug from permuter rename

Permuter improvements (kept as-is):
- Declaration reorders, variable extractions, cast fixes across
  char/, gesture/, hamobj/, rndobj/, ui/, flow/, world/ subsystems
- MILO_NOTIFY -> MILO_LOG, GetPracticeFrames arg order fix
- Comparison equivalence changes (unsigned casts, bitwise AND)

+++ src/system/moviebink/BinkMovieImpl.cpp
@@ -37,8 +37,8 @@ MovieInternalBuffers::MovieInternalBuffers() {
 MovieInternalBuffers::~MovieInternalBuffers() {
     delete mBinks[16];
     mBinks[16] = nullptr;
-    for (int j = 0; j < 2; j++) {
-        for (int i = 0; i < 2; i++) {
+    for (int j = 0; (unsigned int)j < 2; j++) {
+        for (int i = 0; (unsigned int)i < 2; i++) {
             int base = j * 2 + i;
             delete mBinks[base];
             mBinks[base] = nullptr;
